feat(dev-utils,agentic-os): harden GitHub issue lifecycle skills, fix structural drift, document workflow
Backlog review session: promoted real friction to GitHub issues (#460-462),
smoke-tested every script in the new GitHub issue lifecycle skill suite
end-to-end, and fixed everything broken along the way per the self-evolution
"fix forward" rule.

Script fixes:
- gh_issue_create.py: auto-creates missing taxonomy labels before live issue
  creation (previously raised an unhandled traceback)
- gh_issue_prioritize.py: extract_friction_tier() now parses this repo's real
  tier:N-suffix taxonomy format (was silently defaulting every issue to P3)
- update_ecosystem_index.py: filters plugin count to dirs with plugin.yaml
  (was counting __pycache__ and deprecated spec-kitty-plugin as real plugins)

Structural fix (ADR-002/003 hub-and-spoke):
- Moved 16 files (context-bundler persona templates, github-issue-agent
  scripts, github-issue-backlog-agent script) from inside skill directories
  to their plugin-root canonical location, registered in symlinks.json.
  audit_plugin_structure.py: 16 errors -> 0.

Policy updates (.agent/rules/self-evolution-policy.md, both real copies):
- Resolved fixes now require a map-debt.md entry too, not just deferred ones
- New Rule 12: new skill files must land in the plugin root first; both
  audit.py and audit_plugin_structure.py required before any skill/plugin
  edit is considered complete (they catch different failure classes)
- map-debt.md backfilled with 5 RESOLVED entries from this session

Documentation (CLAUDE.md + GEMINI.md/copilot-instructions.md/AGENTS.md
mirrors, kept in sync as full copies per established convention):
- New "Idea Intake Entry Points" section: GitHub Issue vs superpowers
  brainstorming vs exploration-cycle-plugin intake-agent, and the confirmed
  handoff-preparer-agent -> docs/superpowers/specs/ bridge between them
- New rule-file-symlink-awareness note (some rule copies are real symlinks,
  others are independent duplicates -- check before editing multiple)
- New dual-audit requirement section
- github-issue-logging-policy.md referenced (was previously untracked and
  unreferenced anywhere)
- README.md/INSTALL.md ecosystem counts corrected (10 plugins / 133 skills)
- intake-agent.md: stale Spec-Kitty routing reference repointed to Superpowers

Test plan:
- 41/41 pytest passed across github-issue-agent, github-issue-backlog-agent,
  github-issue-prioritizer, issue-worktree-agent, issue-pr-lifecycle-agent
  (39 pre-existing + 2 new regression tests for the label auto-create fix)
- symlink_manager.py diagnose: All links OK
- audit.py + audit_plugin_structure.py: clean on all 4 touched plugins
