+++ CLAUDE.md
@@ -163,6 +163,33 @@ skill, or sub-agent in this repo. Three key capabilities:
 
 ---
 
+## Idea Intake Entry Points
+
+Three front doors exist for a new idea, problem, or need. They are **not 1:1** — pick by how
+well the problem is already understood, not by idea "type":
+
+| Starting point | Entry point | Output |
+|---|---|---|
+| Already know exactly what's broken/needed | `github-issue-agent` — file the issue directly | GitHub Issue |
+| Know WHAT to build, need to design the HOW (single subsystem, one session) | `superpowers:brainstorming` | `docs/superpowers/specs/YYYY-MM-DD-<topic>-design.md` |
+| Don't know the shape yet — unknowns, multiple stakeholders, need a prototype first | `exploration-cycle-plugin`'s `intake-agent` (multi-session discovery) | `exploration/handoffs/handoff-package.md` |
+
+**Confirmed bridge, not a guess:** `handoff-preparer-agent.md` explicitly offers a "Superpowers"
+destination and writes to the *same* `docs/superpowers/specs/` path/format `brainstorming`
+produces directly. Exploration-cycle is a heavier front-end for fuzzy problems — it funnels into
+the same `writing-plans` → `docs/superpowers/plans/` pipeline once the problem is shaped, not a
+competing path. (The "Spec-Kitty" destination also offered there is unused in this repo — see the
+spec-kitty note above.)
+
+**When to create/link the GitHub issue:** at the commit-to-build moment (after `writing-plans`
+produces a plan doc), not at idea time — early design thinking can dead-end, and tracking
+abandoned explorations is noise. If an issue already exists and the fix turns out to need real
+design work, drop into brainstorming from there and link the resulting spec back onto the *same*
+issue (`gh_issue_comment.py`) rather than opening a duplicate — see
+`github-issue-logging-policy.md` §3, Root-Cause Consolidation.
+
+---
+
 **spec-kitty is not installed or used in this repo.** `plugins/spec-kitty-plugin/` is legacy/deprecated
 (superseded by the native Spec Kitty CLI, per README.md) and was never part of the tracked local plugin
 set in `plugin-sources.json`. Do not suggest routing work to spec-kitty or `spk-*` skills unless the user
@@ -347,6 +374,20 @@ python3 .agents/skills/symlink-manager/scripts/symlink_manager.py diagnose
 Fix any BROKEN entries before committing. A broken symlink in `plugins/` will silently fail at install time.
 Shared scripts live in `plugins/<plugin>/scripts/` and are symlinked into each skill's `scripts/` — if you add a new shared script, add it to `symlinks.json` then run `restore`.
 
+### Run both plugin audits after any skill/plugin create or update
+`audit.py` (compliance) and `audit_plugin_structure.py` (structural) check different things — passing
+one does not mean the other passes. A new script or asset file written directly inside a skill directory
+instead of the plugin root (ADR-002/003 hub-and-spoke) is invisible to `audit.py` and only caught by
+`audit_plugin_structure.py`. Run both before considering any new or edited skill/plugin complete:
+
+```bash
+python3 plugins/agent-scaffolders/scripts/audit.py --path plugins/<plugin-name>
+python3 plugins/agent-scaffolders/scripts/audit_plugin_structure.py plugins/<plugin-name>
+```
+
+Fix any structural errors via `symlink_manager.py` (move the real file to the plugin root, add a
+`symlinks.json` entry, `restore`) — never `mv`/`ln -s` by hand. See `self-evolution-policy.md` Rule 12.
+
 ### Scaffolding New Plugins/Skills
 Use these skills rather than hand-rolling structure:
 - `create-plugin` — full plugin scaffold with discovery interview
@@ -356,7 +397,14 @@ Use these skills rather than hand-rolling structure:
 Then run `plugin_add.py` to deploy.
 
 ### Active Rule Files
-Full rule definitions live in `.agent/rules/` — these are the authoritative source, CLAUDE.md carries only the key non-negotiables:
+Full rule definitions live in `.agent/rules/` — these are the authoritative source, CLAUDE.md carries only the key non-negotiables.
+
+**Some rule files exist as multiple logical copies — check before editing more than one.** A few are real
+symlinks (e.g. `plugins/agent-agentic-os/skills/self-evolution/references/self-evolution-policy.md` →
+`plugins/agent-agentic-os/rules/self-evolution-policy.md`; edit the real target once, both update). Others
+are genuinely independent duplicate files with no symlink relationship (e.g. `.agent/rules/self-evolution-policy.md`
+is a separate copy from the plugin's own copy — each needs its own edit). Run `ls -la` / `readlink` on every
+known copy before editing to avoid either double-editing a symlink target or missing an independent duplicate.
 - `coding-conventions.md` — dual-layer docs, file headers, type hints, naming, `tool_inventory.json` registration
 - `dependency-management.md` — pip-compile workflow, no manual pip install, tiered hierarchy
 - `plugin-architecture-policy.md` — decoupling, hub-and-spoke, relative paths, self-contained skills
