+++ plugins/agent-agentic-os/rules/self-evolution-policy.md
@@ -22,7 +22,7 @@ This policy governs how agents must respond when a tool call, subprocess, web au
 
 1. **Verify Edit Boundaries First**: Before making any autonomous edits to repair a failure or friction, check the permitted edit boundaries (e.g., `<project_root>/` or designated module directories). If a repair requires modifying code outside these directories, **escalate to the user immediately**—do not ask for forgiveness.
 2. **Three-Attempt Maximum**: Do not loop or retry a failing repair silently. Attempt to fix a problem up to **three times**. If the third attempt fails, **hard stop** and present the formal Escalation Template to the user with the evidence bundle.
-3. **Update The Map, Not Just the Diary**: Every fix must update the relevant domain playbooks, rules, or reference files (`<project_root>/docs/` or `<project_root>/references/`). A fix that is not recorded is a future regression. Log the evolution step in your history logs with the exact target, action, and outcome.
+3. **Update The Map, Not Just the Diary**: Every fix must update the relevant domain playbooks, rules, or reference files (`<project_root>/docs/` or `<project_root>/references/`). A fix that is not recorded is a future regression. Log the evolution step in your history logs with the exact target, action, and outcome. **This includes fixes applied immediately, not just deferred ones** — add a `Status: RESOLVED` entry to `map-debt.md` for every Tier 0–3 friction event, even when patched in the same turn. A registry with zero entries because everything gets fixed inline is not evidence of a healthy system — it means recurring failure patterns (e.g. the same class of bug breaking multiple scripts) are invisible until someone happens to notice them manually.
 4. **Autonomy & Permission Gates**:
    - **Auto-approved**: Adding new functions/exports, fallback routines/selectors, and appending diffs for modified functions.
    - **Explicit Confirmation Gated**: Renaming or moving files.
@@ -34,6 +34,7 @@ This policy governs how agents must respond when a tool call, subprocess, web au
 9. **Refine Prompt Templates on Ingesting Model Outputs**: Every time you ingest and process responses from external models or APIs, evaluate their quality (checking for lazy placeholders, format violations, or structural errors) and immediately update the template or prompt files to guard against observed deficiencies.
 10. **Synchronize Manifests on Decommissioning/Deletion**: Whenever a skill, file, or plugin is decommissioned, renamed, or deleted, immediately audit and remove its entries from the project's central manifests, registers, and configuration maps (such as `symlinks.json`). Leaving orphaned entries in manifests causes automated scripts (like the symlink manager or plugin installer) to recreate deleted directories as untracked files.
 11. **Pre-Deletion Git History Check — Mandatory**: Before proposing or executing any file deletion, run `git log --follow -- <file>` to understand the file's history. Before assuming a file is dead, orphaned, or replaced — check whether a `.py` (or equivalent) replacement exists, whether symlinks reference it, and whether its purpose is documented in any prior commit message. **Never assume a file is safe to delete because of its extension, name, or apparent disuse.** Confirm with the user if the history reveals any prior active use or replacement migration.
+12. **New Skill Files Must Land in the Hub, Not the Spoke — Mandatory**: When creating a new script, template, or asset file for a skill, write it to the plugin root (`plugins/<plugin>/scripts/`, `plugins/<plugin>/assets/`) first and symlink it into `skills/<skill>/...` via `symlink_manager.py` — never write the real file directly inside a skill directory. This applies even when the file currently has only one consumer skill; sharing potential is not the deciding factor, plugin-root placement is (ADR-002/ADR-003). Run `audit_plugin_structure.py <plugin>` before considering any new skill or script complete — it catches real files living inside skill dirs that the compliance audit (`audit.py`) does not flag as an error. This drift already happened once undetected across 16 files (context-bundler persona templates, the full github-issue-agent script set) before being caught in a manual review — see `map-debt.md`, 2026-07-25.
 
 ---
 
@@ -56,8 +57,8 @@ Agents must not silently work around broken, unclear, missing, or awkward reposi
 Group all failures and workarounds into exactly one tier to determine the required action:
 - **Tier 0 (Friction/Workaround)**: Bypassed an existing capability or used a temporary workaround.
   * *Required Response (pick one)*:
-    - If the fix is small + inside allowed edit boundaries: patch it now, update The Map (rules/docs).
-    - If the fix is not safe or not small: record **Map Debt** in the designated register (e.g., `<project_root>/.agent/map-debt.md` or local rules equivalent) and log the event.
+    - If the fix is small + inside allowed edit boundaries: patch it now, update The Map (rules/docs), **and log a `Status: RESOLVED` entry in `map-debt.md`** — resolved fixes still count.
+    - If the fix is not safe or not small: record **Map Debt** in the designated register (e.g., `<project_root>/.agent/map-debt.md` or local rules equivalent) as `Status: OPEN` and log the event.
     - If friction is repeated/blocking: escalate to the user.
 - **Tier 1 (Gap)**: Capability doesn't exist yet (build missing piece, no evidence needed).
 - **Tier 2 (Failure)**: Code or script exists but is broken or returns errors (patch minimal code, save logs/stack traces).
