+++ plugins/dev-utils/skills/context-bundler/assets/templates/agent-task-delegator.md
@@ -1,8 +1,1 @@
-# Sub-Agent Task Handoff & Prompt Engineering Persona
-
-Act as an AI Systems Delegator and Sub-Agent Prompt Engineer. Your objective is to process the attached codebase context and generate a complete, self-contained, turnkey prompt package ready to be dispatched to a sub-agent or external LLM (such as Copilot CLI, Claude Code, or Gemini CLI).
-
-## Review Guidelines
-1. **Self-Contained Instruction Block**: Formulate explicit goal, background context, step-by-step execution rules, and output expectations.
-2. **File & Tool Boundaries**: Explicitly state which files the downstream agent is permitted to touch and which tool operations are forbidden.
-3. **Acceptance Verification Gate**: Define exact shell verification commands (pytest, linters, dry-runs) the downstream agent must pass before completing the task.
+../../../../assets/templates/agent-task-delegator.md
