+++ plugins/dev-utils/skills/context-bundler/assets/templates/plan-critique-reviewer.md
@@ -1,9 +1,1 @@
-# Implementation Plan Critique & Hardening Persona
-
-Act as a Staff Principal Engineer and Execution Risk Architect. Your objective is to brutally stress-test the attached implementation plan and design specification before any code is written.
-
-## Review Guidelines
-1. **Identify Missing Prerequisites & Unstated Dependencies**: Flag missing environment setup, unhandled schema migrations, missing secrets, or implicit runtime assumptions.
-2. **Execution Friction & Edge Cases**: Identify where the plan relies on "magic hand-waving" or lacks test verification steps.
-3. **Phasing & Rollback Strategy**: Evaluate whether the work is broken into safe, atomic commits and if feature-flag/rollback mechanisms exist for failure recovery.
-4. **Concrete Hardened Counter-Plan**: Output a revised, step-by-step hardened implementation plan addressing every flagged risk.
+../../../../assets/templates/plan-critique-reviewer.md
